backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
werkzeug==3.0.0
python-dotenv==1.0.0
requests==2.31.0
